Added finite linger to zmq base
During use of a flowgraph wrapped in another
Qt GUI, repeated starting and stopping of the
flowgraph would teardown and build a hw source
flowgraph communicating to a signal processing
flowgraph through zmq.  Occasionally (seemingly
randomly) stopping and tearing down the source
flowgraph would freeze the GUI.  Debug traces
suggest the zmq context is hanging while waiting
for unsent messages.

Setting the socket to have a finite linger seems
to fix this behavior.  It is not known for sure
that the infinite linger is the cause of the hangs.