Added CCSDS Reed-Solomon decoder block
A new CCSDS RS decoder FEC-API definition block
was created and the existing code in reed_solomon
subdir was wrapped.  Some input source files
had to be generated, see the reed-solomon/CMakeLists.txt
for details.  The conventional basis version was
wrapped and the dual-basis transformation was
implemented in the hier block in the python
directory using the TAL1TAB and TALTAB basis mappings
exported via swig.  Similar coding pattern could
be used for the rest of the RS implementations and
the encoder definitions too.